+++ src/main/java/org/ahpuh/surf/like/domain/LikeRepository.java
@@ -1,6 +1,5 @@
-package org.ahpuh.surf.like.repository;
+package org.ahpuh.surf.like.domain;
 
-import org.ahpuh.surf.like.entity.Like;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LikeRepository extends JpaRepository<Like, Long> {
